DictionaryRepositories were moved to Repositories/Dictionary directory
